refactor(optim-config): split parsing (reading) from cross-validation
Move validate_optim_config and its _check_* helpers out of parsing.py
into a new validation.py, mirroring the convention of separating
input-reading from business-rule validation. Update all call sites
that imported validate_optim_config directly from .parsing.

Also mark the same mixing pattern with TODOs where it recurs elsewhere
(resolve_components.py, resolve_library.py, study.py) for a future pass.
