+++ src/gems_craft/study/study.py
@@ -53,6 +53,9 @@ class Study:
             mk: components[0].model for mk, components in self.model_components.items()
         }
 
+    # TODO: this is a second, disjoint consistency check alongside
+    # resolve_components.consistency_check() — consider consolidating both
+    # into one validation module for System/Study.
     def check_consistency(self) -> None:
         """Validate that the database supplies data for every parameter of every
         component defined in the system.
